+++ src/Resources/contao/dca/tl_fernschach_spieler.php
@@ -201,7 +201,7 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 	'palettes' => array
 	(
 		'__selector__'                => array('death', 'honor_25', 'honor_40', 'honor_50', 'honor_60', 'honor_70', 'honor_president', 'honor_member', 'sepaBeitrag', 'sepaNenngeld'),
-		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,streichung,patron,memberships,verein;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},klassenberechtigung,turnierAnmeldungenBewerbungen;{iccf_legend:hide},titelinfo;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},beitragsschulden,checkBeitrag,contribution_paid;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info;{publish_legend},published'
+		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,streichung,patron,memberships,verein;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},spielberechtigungen,klassenberechtigung,turnierAnmeldungenBewerbungen;{iccf_legend:hide},titelinfo;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},beitragsschulden,checkBeitrag,contribution_paid;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info;{publish_legend},published'
 	),
 
 	// Subpalettes
@@ -821,6 +821,53 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 				array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'putDate')
 			),
 			'sql'                     => "int(8) unsigned NOT NULL default '0'"
+		),
+		'spielberechtigungen' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['spielberechtigungen'],
+			'exclude'                 => true,
+			'inputType'               => 'multiColumnWizard',
+			'eval'                    => array
+			(
+				'tl_class'            => 'clr',
+				'buttonPos'           => 'top',
+				'columnFields'        => array
+				(
+					'datum' => array
+					(
+						'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['spielberechtigungen_datum'],
+						'exclude'                 => true,
+						'search'                  => false,
+						'sorting'                 => true,
+						'flag'                    => 12,
+						'inputType'               => 'text',
+						'eval'                    => array
+						(
+							'maxlength'           => 10,
+							'style'               => 'width: 200px',
+							'rgxp'                => 'date',
+							'tl_class'            => 'w50 wizard',
+							'datepicker'          => true
+						),
+					),
+					'datei' => array
+					(
+						'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['spielberechtigungen_datei'],
+						'exclude'                 => true,
+						'search'                  => false,
+						'inputType'               => 'fileTree',
+						'eval'                    => array
+						(
+							'multiple'            => false,
+							'fieldType'           => 'checkbox',
+							'files'               => true,
+							'mandatory'           => false,
+							'style'               => 'width: 200px',
+						),
+					),
+				)
+			),
+			'sql'                   => "blob NULL"
 		),
 		'klassenberechtigung' => array
 		(
@@ -1888,7 +1935,7 @@ class tl_fernschach_spieler extends \Backend
 		{
 			switch($session['filter']['tl_fernschach_spielerFilter']['tfs_filter'])
 			{
-				case '1': // Alle Mitglieder
+				case '1': // Nur Mitglieder (Aktiver Mitgliedschaftszeitraum)
 					$objPlayers = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_spieler")
 					                                      ->execute();
 					$arrPlayers = array();