+++ src/Resources/contao/languages/de/tl_module.php
@@ -21,6 +21,7 @@ $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_linkingMembers'] = array(
 $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_tournamentRoot'] = array('Turnierkategorie', 'Oberste Turnierkategorie festlegen. Es werden nur Kategorien angezeigt, bei denen die Checkbox "In Turnieranmeldung anzeigen" aktiviert ist. Ohne Auswahl, werden alle Kategorien verwendet.');
 $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_tournamentText'] = array('Einleitungstext', 'Einleitungstext für das Formular. Beispiel: "Hiermit melde ich mich zu folgendem Fernschachturnier an:"');
 $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_bewerbung'] = array('Bewerbungsformular', 'Das Meldeformular speichert die Daten als Turnierbewerbung.');
+$GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_radio'] = array('Radio-Buttons bei Turnierauswahl', 'Die Turnierauswahl im Meldeformular mit Radio-Buttons statt Select-Liste ausgeben.');
 
 $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_minBuchungen'] = array('Buchungen Minimum', 'Mindestanzahl der Buchungen, die angezeigt werden sollen.');
 $GLOBALS['TL_LANG']['tl_module']['fernschachverwaltung_maxBuchungen'] = array('Buchungen Maximum', 'Maximalanzahl der Buchungen, die angezeigt werden sollen. 0 = alle Buchungen anzeigen.');